python-recognition/requirements.txt
face-recognition==1.3.0
opencv-python==4.7.0.72
numpy==1.24.2
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2 